ld/elf.[ch]: long -> int32

+++ src/cmd/ld/elf.c
@@ -46,8 +46,8 @@ elfstrtab(void)
 }
 
 void
-elf32phdr(void (*putl)(long), ulong type, ulong off, ulong vaddr, ulong paddr,
-	ulong filesz, ulong memsz, ulong prots, ulong align)
+elf32phdr(void (*putl)(int32), uint32 type, uint32 off, uint32 vaddr, uint32 paddr,
+	uint32 filesz, uint32 memsz, uint32 prots, uint32 align)
 {
 	putl(type);
 	putl(off);
@@ -60,9 +60,9 @@ elf32phdr(void (*putl)(long), ulong type, ulong off, ulong vaddr, ulong paddr,
 }
 
 void
-elf32shdr(void (*putl)(long), ulong name, ulong type, ulong flags, ulong vaddr,
-	ulong off, ulong sectsz, ulong link, ulong addnl, ulong align,
-	ulong entsz)
+elf32shdr(void (*putl)(int32), uint32 name, uint32 type, uint32 flags, uint32 vaddr,
+	uint32 off, uint32 sectsz, uint32 link, uint32 addnl, uint32 align,
+	uint32 entsz)
 {
 	putl(name);
 	putl(type);
@@ -77,7 +77,7 @@ elf32shdr(void (*putl)(long), ulong name, ulong type, ulong flags, ulong vaddr,
 }
 
 static void
-elf32sectab(void (*putl)(long))
+elf32sectab(void (*putl)(int32))
 {
 	seek(cout, HEADR+textsize+datsize+symsize, 0);
 	elf32shdr(putl, Stitext, Progbits, Salloc|Sexec, INITTEXT,
@@ -94,8 +94,8 @@ elf32sectab(void (*putl)(long))
 void
 elf32(int mach, int bo, int addpsects, void (*putpsects)(Putl))
 {
-	ulong phydata;
-	void (*putw)(long), (*putl)(long);
+	uint32 phydata;
+	void (*putw)(int32), (*putl)(int32);
 
 	if(bo == ELFDATA2MSB){
 		putw = wput;
@@ -172,8 +172,8 @@ elf32(int mach, int bo, int addpsects, void (*putpsects)(Putl))
  */
 
 void
-elf64phdr(void (*putl)(long), void (*putll)(vlong), ulong type, uvlong off,
-	uvlong vaddr, uvlong paddr, uvlong filesz, uvlong memsz, ulong prots,
+elf64phdr(void (*putl)(int32), void (*putll)(vlong), uint32 type, uvlong off,
+	uvlong vaddr, uvlong paddr, uvlong filesz, uvlong memsz, uint32 prots,
 	uvlong align)
 {
 	putl(type);		
@@ -187,9 +187,9 @@ elf64phdr(void (*putl)(long), void (*putll)(vlong), ulong type, uvlong off,
 }
 
 void
-elf64shdr(void (*putl)(long), void (*putll)(vlong), ulong name, ulong type,
-	uvlong flags, uvlong vaddr, uvlong off, uvlong sectsz, ulong link,
-	ulong addnl, uvlong align, uvlong entsz)
+elf64shdr(void (*putl)(int32), void (*putll)(vlong), uint32 name, uint32 type,
+	uvlong flags, uvlong vaddr, uvlong off, uvlong sectsz, uint32 link,
+	uint32 addnl, uvlong align, uvlong entsz)
 {
 	putl(name);
 	putl(type);
@@ -204,7 +204,7 @@ elf64shdr(void (*putl)(long), void (*putll)(vlong), ulong name, ulong type,
 }
 
 static void
-elf64sectab(void (*putl)(long), void (*putll)(vlong))
+elf64sectab(void (*putl)(int32), void (*putll)(vlong))
 {
 	seek(cout, HEADR+textsize+datsize+symsize, SEEK__START);
 	elf64shdr(putl, putll, Stitext, Progbits, Salloc|Sexec, INITTEXT,
@@ -222,7 +222,7 @@ void
 elf64(int mach, int bo, int addpsects, void (*putpsects)(Putl))
 {
 	uvlong phydata;
-	void (*putw)(long), (*putl)(long);
+	void (*putw)(int32), (*putl)(int32);
 	void (*putll)(vlong);
 
 	if(bo == ELFDATA2MSB){
